+++ buildSrc/src/main/kotlin/plugin/OpenGLPlugin.kt
@@ -6,7 +6,7 @@ import org.gradle.kotlin.dsl.task
 
 class OpenGLPlugin : Plugin<Project> {
     override fun apply(target: Project): Unit = target.run {
-        val generateOpenGLTask = task<GenerateOpenGLTask>("generateOpenGL")
+        val generateOpenGLTask = task<GenerateOpenGLNativeTask>("generateOpenGL")
 
         // TODO: figure out the correct tasks to wire up
